% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{earliest_time}
\alias{earliest_time}
\alias{current_time}
\title{Retrieve useful times in UTC}
\usage{
earliest_time()

current_time()
}
\value{
POSIXct time
}
\description{
\itemize{
\item \code{earliest_time} retrieves a reasonable early timestamp
\item \code{current_time} retrieves the current timestamp
}
}
