% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emolt-package.R
\docType{package}
\name{emolt}
\alias{emolt-package}
\alias{emolt}
\title{emolt - manage local data of DFO-MPO calanus database}
\description{
Access eMOLT data from within R.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/BigelowLab/emolt}
}

}
