% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path.R
\name{emolt_path}
\alias{emolt_path}
\title{Build an emolt path}
\usage{
emolt_path(..., root = emolt_root())
}
\arguments{
\item{...}{components of a path to build}

\item{root}{char, the root filepath}
}
\value{
a fully formed file patha ala file.path
}
\description{
Build an emolt path
}
