% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path.R
\name{emolt_root}
\alias{emolt_root}
\title{Retrieve the eMOLT root path}
\usage{
emolt_root(filename = "~/.emolt")
}
\arguments{
\item{filename}{char, the name of the file that has the eMOLT data path}
}
\value{
char the root path
}
\description{
Retrieve the eMOLT root path
}
