% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch.R
\name{fetch_errdap_csv}
\alias{fetch_errdap_csv}
\alias{fetch_do}
\alias{fetch_temp}
\title{Fetch eMOLT ERRDAP CSV data}
\usage{
fetch_errdap_csv(
  x,
  filename = "emolt_download.csv",
  compress = TRUE,
  overwrite = TRUE,
  timeout = 600,
  ...
)

fetch_do(
  dates = c(earliest_time(), current_time()),
  filename = emolt_path("raw", "emolt_do.csv"),
  ...
)

fetch_temp(
  dates = c(earliest_time(), current_time()),
  filename = emolt_path("raw", "emolt_temp.csv"),
  ...
)
}
\arguments{
\item{x}{char, the URL of the data}

\item{filename}{char, the name of the file to write to}

\item{compress}{logical, if TRUE compres the result}

\item{overwrite}{logical, if the destination \code{filename} already exists, then
set this to TRUE to allow overwrites}

\item{timeout}{numeric, number of seconds before a timeout error}

\item{...}{other arguments for \code{fetch_errdap_csv} and \code{download.file}}

\item{dates}{POSIXct, a two element timestamp vector to select from}
}
\value{
numeric, 0 for success and 1 for anything else
}
\description{
\itemize{
\item \code{fetch_do} fetches eMOLT dissolved oxygen tables
\item \code{fetch_temp} fetches eMOLT temperature tables
}
}
