% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path.R
\name{has_data_path}
\alias{has_data_path}
\title{Test if a data path config file exists}
\usage{
has_data_path(filename = "~/.emolt")
}
\arguments{
\item{filename}{char, the name of the config files}
}
\value{
logical TRUE if the config file exists
}
\description{
Test if a data path config file exists
}
