% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{raw_as_sf}
\alias{raw_as_sf}
\alias{sf_as_raw}
\title{Convert raw data to sf or vice versa}
\usage{
raw_as_sf(x)

sf_as_raw(x)
}
\arguments{
\item{x}{tibble of raw data or an sf object}
}
\value{
sf table or a tibble
}
\description{
Convert raw data to sf or vice versa
}
