% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path.R
\name{set_data_path}
\alias{set_data_path}
\title{Write the path configuration file}
\usage{
set_data_path(path, filename = "~/.emolt")
}
\arguments{
\item{path}{char, the path to the data - like '/mnt/ecocast/coredata/emolt'}

\item{filename}{char, the name of the config files}
}
\description{
Write the path configuration file
}
